move player module to own file
