+++ scrabble.js
@@ -3,19 +3,8 @@ var Scrabble = function(){
   this.players = [];
 };
 
-// PLAYER Constructor
-var Player = function(name) {
-  this.name = name;
-  this.plays = [];
-  this.score = 0;
-};
-
 // SCRABBLE Prototype
 Scrabble.prototype = {
-  //- FUNCTION to add a new player
-  addPlayer: function(name) {
-    this.players.push(new Player(name));
-  },
 
   //- FUNCTION to score a single word
   scoreWord: function(word) {
@@ -71,36 +60,6 @@ Scrabble.prototype = {
   },
 };
 
-// PLAYER Prototype - function for Player to play(word)
-Player.prototype.play = function(word) {
-  Scrabble.scoreWord.call(this, word);
-};
-
-// PLAYER Prototype - function to return totalScore()
-// Player.prototype.totalScore = function() {
-// ;
-// }
-
-// PLAYER Prototype - function to show win (T/F)
-// Player.prototype.hasWon = function() {
-//   if =< 100 points
-// }
-
-// PLAYER Prototype - function to return highest WORD
-// Player.prototype.highestScoringWord = function() {
-// returns one word
-// }
-
-// PLAYER Prototype - function to return highest SCORE
-// Player.prototype.highestWordScore = function() {
-// returns a score
-// }
-
-// TILEBAG Constructor
-// var TileBag = function() {
-//
-// }
-
 module.exports = Scrabble; // starter code
 
 /**********************END OF MODULES****************************/
@@ -112,20 +71,3 @@ var my_game = new Scrabble();
 console.log(my_game.scoreWord("zzz"));
 console.log(my_game.scoreWord("aaaaaaa"));
 console.log(my_game.highestScoreFrom(["aaa", "zzj", "zzx"]));
-
-// print out single player data
-var player1 = my_game.addPlayer("Tamiko");
-var player2 = my_game.addPlayer("Tofu");
-console.log(my_game.players);
-
-console.log(my_game.scoreWord.call(player1, "aaa"));
-
-// console.log(player1.name);
-// console.log(player1.plays);
-// console.log(player1.play("aaa"));
-
-
-// console.log(Object.getPrototypeOf(Scrabble));
-// console.log(tamiko.score);
-
-// console.log(tamiko.plays);
